Mypy zero-error sprint: 409->0 errors across 82 files
Steel Spine Phase 10: Complete type safety enforcement

- Fixed 409 mypy errors across 82 files achieving ZERO remaining errors
- Error categories eliminated:
  * 52+ [union-attr]: Added None guards/assertions for Optional field access
  * 22+ [var-annotated]: Added explicit Dict/List type annotations
  * 19+ [assignment]: Fixed type mismatches and added Optional annotations
  * 14+ [no-redef]: Suppressed try/except import fallback redefinitions
  * 12+ [misc]: Fixed final-name, mixin, and type alias conflicts
  * 11+ [attr-defined]: Handled optional deps and module re-exports
  * 9+ [operator]: Fixed str-used-as-numeric with proper casts/lookups
  * 8+ [dict-item]: Fixed dict type annotations
  * Various: return-value, arg-type, has-type, name-defined, override

- Key refactoring patterns applied:
  * _require_client() helpers for Optional connection guards
  * _ensure_graph() helpers for Optional reasoning graph access
  * Dict[str, AgentInstance] iteration fix (keys vs values)
  * Python 3 except-variable scoping fixes (saved to outer var)
  * AsyncIterator override annotations for streaming backends
  * Optional[Any] for lazy-init polymorphic attributes
  * isinstance guards for Union type narrowing

- Tests: 2668 passed, 0 failed, 33 skipped (unchanged)
- Coverage: 55.52% (unchanged, ratchet: 55%)
- Mypy: 0 errors (was 409 at full-scope scan)

+++ core/iaas/deduplication.py
@@ -107,7 +107,7 @@ class MinHashDeduplicator:
 
     def _jaccard_from_signatures(self, sig1: np.ndarray, sig2: np.ndarray) -> float:
         """Estimate Jaccard similarity from MinHash signatures."""
-        return np.mean(sig1 == sig2)
+        return float(np.mean(sig1 == sig2))
 
     def _lsh_hash(self, signature: np.ndarray, band_idx: int) -> int:
         """Hash a band of the signature for LSH bucketing."""
